# Optimized the 'tobitmask' for Arm64.