1. Added Arm64 implementation of `tobitmask`.
It's Arm64 equivalent implementation for x86 `movemasks`.
The implementation idea is from:
https://community.arm.com/arm-community-blogs/b/infrastructure-solutions-blog/posts/porting-x86-vector-bitmask-optimizations-to-arm-neon

3. The all cases in `velox_exec_test` were passed.
4. `velox_filter_benchmark`: 

Original:
```
[root@c542207d9649 tests]# ./velox_filter_benchmark
WARNING: Benchmark running in DEBUG mode
============================================================================
[...]/velox/type/tests/FilterBenchmark.cpp     relative  time/iter   iters/s
============================================================================
scalarDense                                                 5.58ms    179.33
simdDense                                       51.841%    10.76ms     92.97
scalarSparse                                                1.94ms    514.79
simdSparse                                      80.786%     2.40ms    415.88
```

Optimized:
```
[root@c542207d9649 tests]# ./velox_filter_benchmark
WARNING: Benchmark running in DEBUG mode
============================================================================
[...]/velox/type/tests/FilterBenchmark.cpp     relative  time/iter   iters/s
============================================================================
scalarDense                                                 5.59ms    178.95
simdDense                                       55.537%    10.06ms     99.38
scalarSparse                                                1.94ms    515.24
simdSparse                                      98.249%     1.98ms    506.22

```

5.  This PR firstly provided the Arm64 `tobitmask` API.  And the next step (will be in the other separated PR): 
Adjust `arm64bitmask` caller API in `Hashtable.cpp` to accelerate `HashJoin` performace on Arm64.