@guyuqi 2024-08-28
> The benchmark result does not look very promising. Can you add the hash table change here and show the performance gain on hash table? This may not worth the extra complicity if the gain is similar to what we see in `velox_filter_benchmark`.

Hi @Yuhta  Thanks for your comments. :-)
This New equivalent implementation leverages Arm neon `shrn`:
![image](https://github.com/user-attachments/assets/20af4532-0f9a-48b7-964c-8696af392f0b)
![image](https://github.com/user-attachments/assets/0193ed39-499f-46d2-82b6-5476c079f134)


It is apparently more efficient than the scalar implementation and naive Neon implementation:
Scalar:
```
template <typename T, typename A>
int genericToBitMask(xsimd::batch_bool<T, A> mask) {
  static_assert(mask.size <= 32);
  alignas(A::alignment()) bool tmp[mask.size];
  mask.store_aligned(tmp);
  int ans = 0;
  for (int i = 0; i < mask.size; ++i) {
    ans |= tmp[i] << i;
  }
  return ans;
}
```

naive Neon implementation:
```
  static int toBitMask(xsimd::batch_bool<T, A> mask, const xsimd::neon&) {
    alignas(A::alignment()) static const int8_t kShift[] = {
        -7, -6, -5, -4, -3, -2, -1, 0, -7, -6, -5, -4, -3, -2, -1, 0};
    int8x16_t vshift = vld1q_s8(kShift);
    uint8x16_t vmask = vshlq_u8(vandq_u8(mask, vdupq_n_u8(0x80)), vshift);
    return (vaddv_u8(vget_high_u8(vmask)) << 8) | vaddv_u8(vget_low_u8(vmask));
  }

```


It appears we could achieve a 20% performance improvement in `simdSparse` of `velox_filter_benchmark`.
 I'd like to refine the code based on your comments. 
Additionally, I'll need some time to add hash table change and will update you as soon as possible. Thanks.