@Yuhta 2024-08-30
20% in `Filter::testValues` usually translates to no visible improvement for a query.  The improvement doesn't justify to keep 2 different formats for bitmask.  Can you canonicalize your 4 bits masks to the common ones and see if there is still improvement?